Remove /shutdown endpoint
The earlier version of the vanity server allowed the user to shutdown
the server using a POST request to a special endpoint, but this forces
the server administrator to save the shutdown token from the logs when
the application starts. If this token is lost, then the only way to
shutdown the application is to kill the process and restart it.

This method uses a generic approach in the vanity library, and the
application that uses the library may use any method to trigger the
shutdown. The example server listens for SIGINT and SIGTERM to terminate
the server process.