Add vanity command to run server library
Prior to this change, I was only using the code in the example directory
to run my vanity server. I'd much rather prefer to have an executable
that takes in some configuration and runs with that, rather than having
to change the code everytime.

This change also gets rid of the `VcsProvider` type and enum, since it
serves no real purpose. The `SetProvider` API has changed to use strings
instead.