Use net.Listener instead of port
Prior to this change, the vanity server could only be configured to
listen on a TCP port. While this works, it would be useful to listen on
a Unix socket rather than a TCP port, given that the server is intended
to be run behind a fronting web server such as Nginx.

This commit changes the Listen method to accept a net.Listener. However,
the default is still to listen on TCP port 2369.